Modifications to enable usage of 4.x version of Microsoft.IdentityModel.Clients.ActiveDirectory.
The prior version of the cloud.config used an authentication endpoint that was not compatible with the 4.x ADAL libraries.  Therefore, anyone that updated their nuget packages would get authentication errors.

With this modification developers can again use the latest and greatest of all nuget packages and the solution will still work.