Merged PR 243931: [Developer Samples]: Add try catch block for encryption samples
Here we handle the case when we have cloud gateway.

In case of cloud gateway, when we try to get the gateway information using `gatewayId`, the call throws a `404 Not Found` error. This error indicates the existence of cloud gateway.

In case of on-premises gateway, we get a success response with information like `publicKey`(required for encryption).

So here, we add try catch block to handle the `404` error and continue the flow with cloud gateway.